pyproject.toml
[project]
name = "learn-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.23",
    "langchain-docling>=0.2.0",
    "langchain-openai>=0.3.16",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.1",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
